backend/requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.8
loguru==0.7.3
playwright==1.50.0
pydantic==2.10.6
pydantic-settings==2.8.0
pymongo==4.11.3
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0